+++ src/plone/restapi/serializer/vocabularies.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
+from plone.restapi.batching import HypermediaBatch
 from plone.restapi.interfaces import ISerializeToJson
 from zope.component import adapter
 from zope.component import getMultiAdapter
+from zope.i18n import translate
 from zope.interface import implementer
 from zope.interface import Interface
 from zope.schema.interfaces import ITitledTokenizedTerm
@@ -19,16 +21,44 @@ class SerializeVocabularyToJson(object):
 
     def __call__(self, vocabulary_id):
         vocabulary = self.context
-        serialized_terms = []
+        title = self.request.form.get('title', '')
+        token = self.request.form.get('token', '')
+
+        terms = []
         for term in vocabulary:
+            if title and token:
+                self.request.response.setStatus(400)
+                return dict(error=dict(
+                    type='Invalid parameters',
+                    message='You can not filter by title and token at the same time.')  # noqa
+                )
+
+            if token:
+                if token.lower() != term.token.lower():
+                    continue
+                terms.append(term)
+            else:
+                if title.lower() not in term.title.lower():
+                    continue
+                terms.append(term)
+
+        batch = HypermediaBatch(self.request, terms)
+
+        serialized_terms = []
+        for term in batch:
             serializer = getMultiAdapter((term, self.request),
                                          interface=ISerializeToJson)
-            serialized_terms.append(serializer(vocabulary_id))
+            serialized_terms.append(serializer())
 
-        return {
-            '@id': vocabulary_id,
-            'terms': serialized_terms
+        result = {
+            '@id': batch.canonical_url,
+            'items': serialized_terms,
+            'items_total': batch.items_total,
         }
+        links = batch.links
+        if links:
+            result['batching'] = links
+        return result
 
 
 @implementer(ISerializeToJson)
@@ -39,12 +69,11 @@ class SerializeTermToJson(object):
         self.context = context
         self.request = request
 
-    def __call__(self, vocabulary_id):
+    def __call__(self):
         term = self.context
         token = term.token
         title = term.title if ITitledTokenizedTerm.providedBy(term) else token
         return {
-            '@id': '{}/{}'.format(vocabulary_id, token),
             'token': token,
-            'title': title
+            'title': translate(title, context=self.request)
         }