+++ src/plone/restapi/tests/test_dxfield_serializer.py
@@ -90,10 +90,15 @@ class TestDexterityFieldSerializing(TestCase):
         self.assertTrue(isinstance(value, six.text_type), 'Not an <unicode>')
         self.assertEqual(u'\xe4\xf6\xfc', value)
 
-    def test_choice_field_serialization_returns_vocabulary_item(self):
+    def test_choice_field_serialization_returns_vocabulary_term(self):
         value = self.serialize('test_choice_field', u'foo')
-        self.assertTrue(isinstance(value, six.text_type), 'Not an <unicode>')
-        self.assertEqual(u'foo', value)
+        self.assertTrue(isinstance(value, dict))
+        self.assertEqual({u'token': u'foo', u'title': None}, value)
+
+    def test_choice_field_with_vocabulary_serialization_returns_vocabulary_term(self):  # noqa
+        value = self.serialize('test_choice_field_with_vocabulary', u'value1')
+        self.assertTrue(isinstance(value, dict))
+        self.assertEqual({u'token': u'token1', u'title': u'title1'}, value)
 
     def test_date_field_serialization_returns_unicode(self):
         value = self.serialize('test_date_field', date(2015, 7, 15))
@@ -136,6 +141,18 @@ class TestDexterityFieldSerializing(TestCase):
         value = self.serialize('test_list_field', [1, 'two', 3])
         self.assertTrue(isinstance(value, list), 'Not a <list>')
         self.assertEqual([1, u'two', 3], value)
+
+    def test_list_field_with_vocabulary_choice_serialization_returns_terms(
+            self):
+        value = self.serialize('test_list_field_with_choice_with_vocabulary',
+                               [u'value1', u'value3'])
+        self.assertTrue(isinstance(value, list), 'Not a <list>')
+        self.assertEqual(
+            [
+                {u'token': u'token1', u'title': u'title1'},
+                {u'token': u'token3', u'title': u'title3'},
+            ],
+            value)
 
     def test_set_field_serialization_returns_list(self):
         value = self.serialize('test_set_field', set(['a', 'b', 'c']))